+++ src/js/views/LottoMachineView.js
